feat: add getResponseOutputText helper for streaming responses
Adds a simple helper function to extract aggregated text output from
ResponseObject instances.

This replaces the automatic output_text property approach from PR #42
with an explicit helper function, providing a cleaner API without
prototype patching.

Usage:
  import { getResponseOutputText } from 'llama-stack-client';

  const stream = await client.responses.create({ stream: true, ... });
  for await (const chunk of stream) {
    if (chunk.type === 'response.completed') {
      const text = getResponseOutputText(chunk.response);
    }
  }

  // Also works for non-streaming responses
  const response = await client.responses.create({ stream: false, ... });
  const text = getResponseOutputText(response);
